typesafe env with zod
