+++ lib/env.ts
@@ -1,0 +1,9 @@
+import { z } from "zod";
+
+const EnvSchema = z.object({
+  NODE_ENV: z.string(),
+});
+
+export type EnvSchema = z.infer<typeof EnvSchema>;
+
+export default EnvSchema.parse(process.env);
