+++ nuxt.config.ts
@@ -1,5 +1,5 @@
 import tailwindcss from "@tailwindcss/vite";
-
+import "./lib/env";
 // https://nuxt.com/docs/api/configuration/nuxt-config
 export default defineNuxtConfig({
   compatibilityDate: "2025-05-15",
