+++ package.json
@@ -15,6 +15,7 @@
     "nuxt": "^3.17.6",
     "tailwindcss": "^4.1.11",
     "vue": "^3.5.17",
-    "vue-router": "^4.5.1"
+    "vue-router": "^4.5.1",
+    "zod": "^3.25.74"
   }
 }
